Added zip library and fixed and improved a lot of core files and update server

+++ Client-Core/src/Core/Crypto.h
@@ -10,7 +10,7 @@ namespace Core
 
 		struct key_t
 		{
-			uint64 Size = 0;
+			uint32 Size = 0;
 			Byte Data[4096];
 		};
